* 'development' of git.qvik.fi:homehapp/homehapp-web-backend:
  fixed staging host
  Use new database hosts
  Added initial home update API route
  New documentation about Deployment management
  Streamlined the container creation arguments
  New Deployment scripts and configs

+++ support/updateCluster.sh
@@ -4,32 +4,26 @@
 REV=`git rev-parse --short HEAD`
 CWD=`pwd`
 CONTAINER_REGISTRY_HOST=eu.gcr.io
-PNAME=$1
-ENV=$2
+ENV=$1
+PNAME="$2-$ENV"
+PBNAME="$2"
+CLUSTER_NAME="homehapp-$ENV"
+CLUSTER_GOOGLE_NAME=""
 
 function printUsage() {
   echo "Required environment variables:"
   echo "  PROJECT_ID:     Google Project ID"
   echo ""
-  echo "Usage PROJECT_ID=id ./support/updateCluster.sh [project_name] [stg|prod]"
+  echo "Usage PROJECT_ID=id ./support/updateCluster.sh [stg,prod] [project_name]"
+  echo "Add -d for dry-run"
 }
 function printUsageAndExit() {
   printUsage
   exit
 }
 
-function exitWithError() {
-    echo "$1" 1>&2
-    exit 1
-}
-
 if [ "$PROJECT_ID" = "" ]; then
   echo "No Google Project ID defined!"
-  printUsageAndExit
-fi
-
-if [ "$PNAME" = "" ]; then
-  echo "No project defined!"
   printUsageAndExit
 fi
 
@@ -38,39 +32,87 @@ if [ "$ENV" = "" ]; then
   printUsageAndExit
 fi
 
-function configGCloud() {
-  local CLUSTER_NAME=`gcloud beta container clusters list --project $PROJECT_ID | awk '{print $1}' | grep "$PNAME-$ENV"`
-  echo "Executing: 'gcloud config set container/cluster $CLUSTER_NAME'"
-  gcloud config set container/cluster $CLUSTER_NAME
+if [ "$PNAME" = "" ]; then
+  echo "No project defined!"
+  printUsageAndExit
+fi
 
-  local CLUSTER_GOOGLE_NAME=`kubectl config view | awk '{print $2}' | grep $CLUSTER_NAME | tail -n 1`
-  echo "Execute: 'kubectl config use-context $CLUSTER_GOOGLE_NAME'"
-  kubectl config use-context $CLUSTER_GOOGLE_NAME
+function setContext() {
+  if [ "$CLUSTER_GOOGLE_NAME" = "" ]; then
+    CLUSTER_GOOGLE_NAME=`kubectl config view | awk '{print $2}' | grep $CLUSTER_NAME | tail -n 1`
+  fi
+
+  if [ "$1" = "-d" ]; then
+    echo "Execute: 'kubectl config use-context $CLUSTER_GOOGLE_NAME'"
+  else
+    kubectl config use-context $CLUSTER_GOOGLE_NAME
+  fi
 }
 
 function updateCluster() {
-  local SOURCE_CONFIG="$CWD/infrastructure/configs/project-controller-tpl.json"
+  local SOURCE_CONFIG="$CWD/infrastructure/configs/$ENV/project-controller-tpl.json"
   local TARGET_CONFIG="$CWD/tmp/$PNAME-controller.json"
   local TMP_FILE="/tmp/$PNAME-controller.json"
   local CURRENT_CONTROLLER=`kubectl get rc | grep $PROJECT_ID | awk '{print $1}' | grep "$PNAME-controller"`
 
   sed "s/:PROJECT_NAME/$PNAME/g" $SOURCE_CONFIG > $TARGET_CONFIG
+  sed "s/:PROJECT_BASE_NAME/$PBNAME/g" $TARGET_CONFIG > $TMP_FILE && mv $TMP_FILE $TARGET_CONFIG
   sed "s/:PROJECT_ID/$PROJECT_ID/g" $TARGET_CONFIG > $TMP_FILE && mv $TMP_FILE $TARGET_CONFIG
   sed "s/:CONTROLLER_NAME/$PNAME-controller-$REV/g" $TARGET_CONFIG > $TMP_FILE && mv $TMP_FILE $TARGET_CONFIG
   sed "s/:REV/$REV/g" $TARGET_CONFIG > $TMP_FILE && mv $TMP_FILE $TARGET_CONFIG
   sed "s/:ENV/$ENV/g" $TARGET_CONFIG > $TMP_FILE && mv $TMP_FILE $TARGET_CONFIG
 
-  echo "Executing: 'kubectl rolling-update $CURRENT_CONTROLLER -f $CWD/tmp/$PNAME-controller.json' --update-period='2m0s'"
-  kubectl rolling-update $CURRENT_CONTROLLER -f "$CWD/tmp/$PNAME-controller.json" || exitWithError "Error running rolling-update"
-  # --update-period="2m0s"
-  rm $TARGET_CONFIG
+  if [ "$1" = "-d" ]; then
+    echo "Executing: 'kubectl rolling-update $CURRENT_CONTROLLER -f $CWD/tmp/$PNAME-controller.json'"
+  else
+    kubectl rolling-update $CURRENT_CONTROLLER -f "$CWD/tmp/$PNAME-controller.json"
+    rm $TARGET_CONFIG
+  fi
 }
 
-configGCloud
+function updateProxyCluster() {
+  local SOURCE_CONFIG="$CWD/infrastructure/configs/$ENV/proxy-controller-tpl.json"
+  local TARGET_CONFIG="$CWD/tmp/$PNAME-proxy-controller.json"
+  local TMP_FILE="/tmp/$PNAME-proxy-controller.json"
+  local UC_PNAME=`echo "${PBNAME}_${ENV}" | awk '{print toupper($0)}'`
+  local CURRENT_CONTROLLER=`kubectl get rc | grep $PROJECT_ID | awk '{print $1}' | grep "$PNAME-proxy-controller"`
+
+  local SERVICE_HOST="${UC_PNAME}_SERVICE_HOST"
+  local SERVICE_PORT="${UC_PNAME}_SERVICE_PORT_ENDPOINT"
+
+  sed "s/:PROJECT_NAME/$PNAME-proxy/g" $SOURCE_CONFIG > $TARGET_CONFIG
+  sed "s/:PROJECT_ID/$PROJECT_ID/g" $TARGET_CONFIG > $TMP_FILE && mv $TMP_FILE $TARGET_CONFIG
+  sed "s/:CONTROLLER_NAME/$PNAME-proxy-controller-$REV/g" $TARGET_CONFIG > $TMP_FILE && mv $TMP_FILE $TARGET_CONFIG
+  sed "s/:REV/$REV/g" $TARGET_CONFIG > $TMP_FILE && mv $TMP_FILE $TARGET_CONFIG
+  sed "s/:ENV/$ENV/g" $TARGET_CONFIG > $TMP_FILE && mv $TMP_FILE $TARGET_CONFIG
+  sed "s/:SERVICE_HOST/$SERVICE_HOST/g" $TARGET_CONFIG > $TMP_FILE && mv $TMP_FILE $TARGET_CONFIG
+  sed "s/:SERVICE_PORT/$SERVICE_PORT/g" $TARGET_CONFIG > $TMP_FILE && mv $TMP_FILE $TARGET_CONFIG
+
+  if [ "$1" = "-d" ]; then
+    echo "Executing: 'kubectl rolling-update $CURRENT_CONTROLLER -f $CWD/tmp/$PNAME-proxy-controller.json'"
+  else
+    kubectl rolling-update $CURRENT_CONTROLLER -f "$CWD/tmp/$PNAME-proxy-controller.json"
+    rm $TARGET_CONFIG
+  fi
+}
+
+if [ "$3" = "-d" ]; then
+  echo "In Dry-Run mode"
+fi
+
+setContext $3
 
 echo ""
-echo "Updating Cluster Pods for project '$PNAME' to revision $REV for environment $ENV"
+echo "Updating Cluster Pods for project '$PBNAME' to revision $REV for environment $ENV"
 
-updateCluster
+updateCluster $3
+
+if [ "$ENV" = "prod" ]; then
+  echo ""
+  echo "Updating Proxy Cluster"
+
+  updateProxyCluster $3
+fi
 
 echo ""
+echo "Finished"
